Code reviewing

+++ Vector2.h
@@ -5,7 +5,7 @@ typedef struct {
 	float y = 0;
 }Vector2;
 
-struct Vector2* Vector2_new(int x = 0, int y = 0){
+struct Vector2* Vector2_new(float x = 0, float y = 0){
 	struct Vector2* result = (struct Vector2*) malloc(sizeof(struct Vector2));
 	
 	Vector2->x = 0;
@@ -25,7 +25,7 @@ void Vector2_add(struct Vector2* vec1, struct Vector2* vec2) {
 	(vec1->x) += (vec2->x);
 	(vec1->y) += (vec2->y);
 }
-void Vector2_add(struct Vector2* vec, int x, int y) {
+void Vector2_add(struct Vector2* vec, float x, float y) {
 	(vec->x) += (x);
 	(vec->x) += (y);
 }
@@ -33,7 +33,7 @@ void Vector2_sub(struct Vector2* vec1, struct Vector2* vec2) {
 	(vec1->x) -= (vec2->x);
 	(vec1->y) -= (vec2->y);
 }
-void Vector2_sub(struct Vector2* vec, int x, int y) {
+void Vector2_sub(struct Vector2* vec, float x, float y) {
 	(vec->x) -= (x);
 	(vec->x) -= (y);
 }
@@ -41,7 +41,7 @@ void Vector2_multi(struct Vector2* vec1, struct Vector2* vec2) {
 	(vec1->x) *= (vec2->x);
 	(vec1->y) *= (vec2->y);
 }
-void Vector2_multi(struct Vector2* vec, int x, int y) {
+void Vector2_multi(struct Vector2* vec, float x, float y) {
 	(vec->x) *= (x);
 	(vec->x) *= (y);
 }
@@ -49,7 +49,7 @@ void Vector2_div(struct Vector2* vec1, struct Vector2* vec2) {
 	(vec1->x) /= (vec2->x);
 	(vec1->y) /= (vec2->y);
 }
-void Vector2_div(struct Vector2* vec, int x, int y) {
+void Vector2_div(struct Vector2* vec, float x, float y) {
 	(vec->x) /= (x);
 	(vec->x) /= (y);
 }
